+++ src/main/java/nextstep/ladder/service/LadderGame.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.service;
 
 public class LadderGame {
   public void start() {
