+++ src/test/java/nextstep/ladder/service/LadderGameTest.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.service;
 
 
 public class LadderGameTest {
